Create Common Pie Chart Component (#2199)
