+++ src/components/cms/SkillUsageCard/SkillUsageCard.js
@@ -11,9 +11,6 @@ import {
   YAxis,
   Tooltip,
   Legend,
-  PieChart,
-  Pie,
-  Sector,
   Cell,
   ResponsiveContainer,
 } from 'recharts';
@@ -24,93 +21,7 @@ import Paper from '@material-ui/core/Paper';
 
 // Static assets
 import './SkillUsageCard.css';
-
-const renderActiveShape = props => {
-  const RADIAN = Math.PI / 180;
-  const {
-    cx,
-    cy,
-    midAngle,
-    innerRadius,
-    outerRadius,
-    startAngle,
-    endAngle,
-    fill,
-    percent,
-    value,
-    name,
-  } = props;
-  const sin = Math.sin(-RADIAN * midAngle);
-  const cos = Math.cos(-RADIAN * midAngle);
-  const sx = cx + (outerRadius + 10) * cos;
-  const sy = cy + (outerRadius + 10) * sin;
-  const mx = cx + (outerRadius + 30) * cos;
-  const my = cy + (outerRadius + 30) * sin;
-  const ex = mx + (cos >= 0 ? 1 : -1) * 22;
-  const ey = my;
-  const textAnchor = cos >= 0 ? 'start' : 'end';
-
-  return (
-    <g>
-      <text x={cx} y={cy} dy={8} textAnchor="middle" fill={fill}>
-        {name}
-      </text>
-      <Sector
-        cx={cx}
-        cy={cy}
-        innerRadius={innerRadius}
-        outerRadius={outerRadius}
-        startAngle={startAngle}
-        endAngle={endAngle}
-        fill={fill}
-      />
-      <Sector
-        cx={cx}
-        cy={cy}
-        startAngle={startAngle}
-        endAngle={endAngle}
-        innerRadius={outerRadius + 6}
-        outerRadius={outerRadius + 10}
-        fill={fill}
-      />
-      <path
-        d={`M${sx},${sy}L${mx},${my}L${ex},${ey}`}
-        stroke={fill}
-        fill="none"
-      />
-      <circle cx={ex} cy={ey} r={2} fill={fill} stroke="none" />
-      <text
-        x={ex + (cos >= 0 ? 1 : -1) * 12}
-        y={ey}
-        textAnchor={textAnchor}
-        fill="#333"
-      >{`${name}: ${value}`}</text>
-      <text
-        x={ex + (cos >= 0 ? 1 : -1) * 12}
-        y={ey}
-        dy={18}
-        textAnchor={textAnchor}
-        fill="#999"
-      >
-        {`(${(percent * 100).toFixed(2)}%)`}
-      </text>
-    </g>
-  );
-};
-
-renderActiveShape.propTypes = {
-  cx: PropTypes.number,
-  cy: PropTypes.number,
-  midAngle: PropTypes.number,
-  innerRadius: PropTypes.number,
-  outerRadius: PropTypes.number,
-  startAngle: PropTypes.number,
-  endAngle: PropTypes.number,
-  fill: PropTypes.string,
-  percent: PropTypes.number,
-  value: PropTypes.number,
-  name: PropTypes.string,
-};
+import PieChartContainer from '../../shared/PieChartContainer';
 
 class SkillUsageCard extends Component {
   static propTypes = {
@@ -123,7 +34,6 @@ class SkillUsageCard extends Component {
     super(props);
     this.state = {
       width: 0,
-      activePieIndex: 0,
     };
   }
 
@@ -139,12 +49,6 @@ class SkillUsageCard extends Component {
   updateWindowDimensions = () => {
     this.setState({
       width: window.innerWidth * 0.8 > 800 ? 800 : window.innerWidth * 0.8,
-    });
-  };
-
-  onPieEnter = (data, index) => {
-    this.setState({
-      activePieIndex: index,
     });
   };
 
@@ -167,10 +71,8 @@ class SkillUsageCard extends Component {
       deviceWiseSkillUsage,
       countryWiseSkillUsage,
     } = this.props;
-    const { width, activePieIndex } = this.state;
-
+    const { width } = this.state;
     let totalSkillUsage = this.totalUsage(dateWiseSkillUsage);
-
     return (
       <div>
         <Paper className="margin-b-md margin-t-md">
@@ -222,28 +124,12 @@ class SkillUsageCard extends Component {
               <div className="sub-title">Device wise Usage</div>
               {deviceWiseSkillUsage && deviceWiseSkillUsage.length ? (
                 <div className="pie-chart">
-                  <ResponsiveContainer width={600} height={350}>
-                    <PieChart>
-                      <Pie
-                        activeIndex={activePieIndex}
-                        activeShape={renderActiveShape}
-                        data={deviceWiseSkillUsage}
-                        cx={300}
-                        cy={175}
-                        innerRadius={80}
-                        nameKey="deviceType"
-                        dataKey="count"
-                        outerRadius={120}
-                        fill="#8884d8"
-                        onMouseEnter={this.onPieEnter}
-                      >
-                        {deviceWiseSkillUsage.map((entry, index) => (
-                          <Cell key={index} fill={entry.color} />
-                        ))}
-                      </Pie>
-                      <Legend wrapperStyle={{ position: 'relative' }} />
-                    </PieChart>
-                  </ResponsiveContainer>
+                  <PieChartContainer
+                    cellData={deviceWiseSkillUsage.map((entry, index) => (
+                      <Cell key={index} fill={entry.color} />
+                    ))}
+                    data={deviceWiseSkillUsage}
+                  />
                 </div>
               ) : (
                 <div className="default-message">
